moved auth context to context folder
