+++ frontend/src/context/auth-context.tsx
@@ -1,0 +1,12 @@
+import { createContext} from "react";
+
+
+interface AuthContextType {
+  accessToken: string | null;
+  setAccessToken: (token: string | null) => void;
+  refreshToken: () => Promise<string | null>;
+  logout: () => void;
+}
+
+export const AuthContext = createContext<AuthContextType | null>(null);
+
